Refactor: Creacion del nuevo modulo 'autenticacion' e implementacion del submodulo 'auth-token'
