+++ src/app/(modulos)/(autenticacion)/auth-token/(presentation)/(pages)/page.tsx
@@ -1,5 +1,5 @@
-import { Metadata } from "next";
-import AuthToken from "@/components/AuthToken/AuthToken";
+﻿import { Metadata } from "next";
+import AuthToken from "../components/AuthToken";
 
 export const metadata: Metadata = {
   title: "2FA Boston ERP",
